+++ CHANGELOG.md
@@ -20,7 +20,9 @@ All notable changes to this project will be documented in this file.
   - Improve how snaphots are created
 - bootloader: default title is the pretty_name.
 - bootloader: Trust only the UKI's uname as kerver when managing UKIs.
+- bootloader: Remove uki from efi dir only if not used anymore.
 - sdboot: get version_id and id to fill the conf entry file correctly.
+- grub2: refer to the uki_name_id when removing entry.
 - common: improves function to get data from uki file (uname, pretty_name).
 
 ### Fixed